fin_research_project/pyproject.toml
[project]
name = "fin_research_project"
version = "0.1.0"
description = "fin_research_project using crewAI"
authors = [{ name = "Your Name", email = "you@example.com" }]
requires-python = ">=3.10,<3.14"
dependencies = [
    "crewai[tools]>=0.150.0,<1.0.0",
    "requests",
    "google-generativeai",
    "fitz",
    "beautifulsoup4",
    "sentence-transformers",
    "faiss-cpu",
    "tqdm"
]

[project.scripts]
fin_research_project = "fin_research_project.main:run"
run_crew = "fin_research_project.main:run"
train = "fin_research_project.main:train"
replay = "fin_research_project.main:replay"
test = "fin_research_project.main:test"
quantitative_research = "fin_research_project.quantitative_main:main"
run_quantitative = "fin_research_project.quantitative_main:run_quantitative_research"

[build-system]
requires = ["hatchling"]
build-backend = "hatchling.build"

[tool.crewai]
type = "crew"
